Replace race PNG images with SVG images
Add logic to scale the and position the icons in the SVG correctly based
on image width/height instead of hardcoding it.

The SC2 race icons are from https://github.com/nephest/sc2-icons, and
are MIT licensed.  The LICENSE for this project was updated to include
the license from the above project.