Use SVG region images for high-DPI displays
The geographical region icons are from
https://github.com/yammadev/flag-icons, and are MIT licensed.  The
LICENSE for this project was updated to include the license from the
above project.

Since flags have a lot of small detail, using the SVG icons universally
do not render well on low-DPI displays. Instead use the 'srcset'
attribute to select the SVG images for high-DPI displays.