Fix tooltip positioning
This is a follow on to commit 85b68843e9d0bfc7417b890f9e3b069e5ed80153.
'this.canvas.width' (and 'this.canvas.height') can't be used for any
sizing/positioning calculations since the canvas is scaled by the device
pixel ratio.

Use 'this.width' (and 'this.height') instead, as these correspond to the
logical size of the canvas. Additionally introduce 'this.fullWidth' (and
'this.fullHeight') for convenience: these include the dimensions of the
edges.